Fix file upload bug
This fixes a bug where file upload fields are "empty" in the form when going back to the form from the confirmation page. The bug is related to the "uid" property getting lost during the validation/deserialization process. The fix involves setting "unknown" to "preserve" in the colander Mapping passed to the SchemaNode.

Requires https://github.com/stefanofontanelli/ColanderAlchemy/pull/78.